requirements.txt
django == 3.1.5
django-crispy-forms
pillow
django-rest-framework
social-auth-app-django
django-pwa